Added new paths for branches Added tests for paths Moved test classes into correct path
