+++ src/test/java/com/lordmat/githubstream/data/GitHubCallerTest.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.lordmat.githubstream.api;
+package com.lordmat.githubstream.data;
 
 import com.lordmat.githubstream.StartManager;
 import com.lordmat.githubstream.data.GitHubCaller;
